+++ packages/lint-suite/src/lib/prettier.ts
@@ -1,0 +1,4 @@
+import eslintConfigPrettier from 'eslint-config-prettier';
+import { defineConfig } from 'eslint/config';
+
+export const prettier = defineConfig([eslintConfigPrettier]);