+++ packages/lint-suite/src/lib/storybook.ts
@@ -5,6 +5,10 @@ export const storybook = defineConfig([
   {
     ignores: ['!.storybook'],
     files: ['**/*.stories.ts'],
-    extends: [storybookEslint.configs['flat/recommended']]
+    extends: [storybookEslint.configs['flat/recommended']],
+    rules: {
+      'storybook/csf-component': 'warn', // Ensure component property is set
+      'storybook/no-stories-of': 'error' // Prevent deprecated storiesOf API
+    }
   }
 ]);